+++ doc/ReleaseNotes.md
@@ -8,6 +8,7 @@ These release notes start with the October 18, 2021 CloudComPy release.
  - improve C2C split distances (issue #56), see test009.py
  - get access to sensor TSL info (sensors position, transformation...) (issue #57), see test041.py
  - get access to E57 structure (issue #58), see test041.py
+ - compute scattering angles (issue #59), see test041.py
  
 ## June 22, 2022 CloudComPy release: